+++ src/components/Model.js
@@ -6,7 +6,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   changeContent, changeType, checkDirection, expandItem, updateRestrictedDropId,
-} from './redux';
+} from '../redux';
 
 const Model = ({
   item,
